@@ -4,7 +4,7 @@
   "private": true,
   "description": "",
   "templateInfo": {
-    "name": "default",
+    "name": "vue3-vuex",
     "typescript": false,
     "css": "sass"
   },
@@ -32,25 +32,25 @@
     "ios >= 8"
   ],
   "author": "",
+  "license": "MIT",
   "dependencies": {
     "@babel/runtime": "^7.7.7",
     "@tarojs/components": "3.0.8",
     "@tarojs/runtime": "3.0.8",
     "@tarojs/taro": "3.0.8",
-    "vue": "^2.6.12",
-    "vue-template-compiler": "^2.6.12",
-    "vuex": "^3.5.1"
+    "vue": "^3.0.0-beta.15",
+    "vuex": "^4.0.0-beta.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.4",
+    "@babel/core": "^7.8.0",
     "@tarojs/mini-runner": "3.0.8",
     "@tarojs/webpack-runner": "3.0.8",
     "@types/webpack-env": "^1.13.6",
+    "@vue/compiler-sfc": "^3.0.0-beta.15",
     "babel-preset-taro": "3.0.8",
     "eslint": "^7.7.0",
     "eslint-config-taro": "3.0.8",
-    "eslint-plugin-vue": "^6.x",
-    "stylelint": "13.6.1",
-    "vue-loader": "^15.9.3"
+    "eslint-plugin-vue": "^7.0.0-alpha.6",
+    "vue-loader": "^16.0.0-beta.3"
   }
 }
